Type resolver refactoring.
Refactored type resolvers like `Resolved` and `PascalCase`.

Also supported `RegExp` as one of native classes.
