+++ src/Resolved.ts
@@ -1,3 +1,7 @@
+import { IsTuple } from "./typings/IsTuple";
+import { NativeClass } from "./typings/NativeClass";
+import { ValueOf } from "./typings/ValueOf";
+
 /**
  * Resolved type erased every methods.
  *
@@ -52,24 +56,7 @@ type ResolvedObject<T extends object> =
         ? Map<ResolvedMain<K>, ResolvedMain<V>>
         : T extends WeakSet<any> | WeakMap<any, any>
           ? never
-          : T extends
-                | Date
-                | Uint8Array
-                | Uint8ClampedArray
-                | Uint16Array
-                | Uint32Array
-                | BigUint64Array
-                | Int8Array
-                | Int16Array
-                | Int32Array
-                | BigInt64Array
-                | Float32Array
-                | Float64Array
-                | ArrayBuffer
-                | SharedArrayBuffer
-                | DataView
-                | Blob
-                | File
+          : T extends NativeClass
             ? T
             : {
                 [P in keyof T]: ResolvedMain<T[P]>;
@@ -86,34 +73,3 @@ type ResolvedTuple<T extends readonly any[]> = T extends []
         : T extends [(infer F)?, ...infer Rest extends readonly any[]]
           ? [ResolvedMain<F>?, ...ResolvedTuple<Rest>]
           : [];
-
-type IsTuple<T extends readonly any[] | { length: number }> = [T] extends [
-  never,
-]
-  ? false
-  : T extends readonly any[]
-    ? number extends T["length"]
-      ? false
-      : true
-    : false;
-
-type ValueOf<Instance> =
-  IsValueOf<Instance, Boolean> extends true
-    ? boolean
-    : IsValueOf<Instance, Number> extends true
-      ? number
-      : IsValueOf<Instance, String> extends true
-        ? string
-        : Instance;
-
-type IsValueOf<Instance, Object extends IValueOf<any>> = Instance extends Object
-  ? Object extends IValueOf<infer Primitive>
-    ? Instance extends Primitive
-      ? false
-      : true // not Primitive, but Object
-    : false // cannot be
-  : false;
-
-interface IValueOf<T> {
-  valueOf(): T;
-}
